# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Tencent_WeKnora_internal_errors.AppError:
    properties:
      code:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_errors.ErrorCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2100
    - 2101
    - 2102
    - 2103
    type: integer
    x-enum-varnames:
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrServiceUnavailable
    - ErrTimeout
    - ErrValidation
    - ErrTenantNotFound
    - ErrTenantAlreadyExists
    - ErrTenantInactive
    - ErrTenantNameRequired
    - ErrTenantInvalidStatus
    - ErrAgentMissingThinkingModel
    - ErrAgentMissingAllowedTools
    - ErrAgentInvalidMaxIterations
    - ErrAgentInvalidTemperature
...
  github_com_Tencent_WeKnora_internal_types.EmbeddingParameters:
    properties:
      dimension:
        type: integer
      truncate_prompt_tokens:
        type: integer
    type: object
...
  github_com_Tencent_WeKnora_internal_types.ModelParameters:
    properties:
      api_key:
        type: string
      base_url:
        type: string
      embedding_parameters:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.EmbeddingParameters'
      interface_type:
        type: string
      parameter_size:
        description: Ollama model parameter size (e.g., "7B", "13B", "70B")
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_types.ModelSource:
    enum:
    - local
    - remote
    - aliyun
    type: string
    x-enum-comments:
      ModelSourceAliyun: Aliyun DashScope model
      ModelSourceLocal: Local model
      ModelSourceRemote: Remote model
    x-enum-descriptions:
    - Local model
    - Remote model
    - Aliyun DashScope model
    x-enum-varnames:
    - ModelSourceLocal
    - ModelSourceRemote
    - ModelSourceAliyun
  github_com_Tencent_WeKnora_internal_types.ModelType:
    enum:
    - Embedding
    - Rerank
    - KnowledgeQA
    - VLLM
    type: string
    x-enum-comments:
      ModelTypeEmbedding: Embedding model
      ModelTypeKnowledgeQA: KnowledgeQA model
      ModelTypeRerank: Rerank model
      ModelTypeVLLM: VLLM model
    x-enum-descriptions:
    - Embedding model
    - Rerank model
    - KnowledgeQA model
    - VLLM model
    x-enum-varnames:
    - ModelTypeEmbedding
    - ModelTypeRerank
    - ModelTypeKnowledgeQA
    - ModelTypeVLLM
...
  internal_handler.CreateModelRequest:
    properties:
      description:
        type: string
      name:
        type: string
      parameters:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.ModelParameters'
      source:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.ModelSource'
      type:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.ModelType'
    required:
    - name
    - parameters
    - source
    - type
    type: object
...
  internal_handler.UpdateModelRequest:
    properties:
      description:
        type: string
      name:
        type: string
      parameters:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.ModelParameters'
      source:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.ModelSource'
      type:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.ModelType'
    type: object
...
info:
  contact:
    name: WeKnora Github
    url: https://github.com/aiplusall/aiplusall-kb
  description: WeKnora 知识库管理系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: WeKnora API
  version: "1.0"
paths:
...
  /messages/{session_id}/{id}:
    delete:
      consumes:
      - application/json
      description: 从会话中删除指定消息
      parameters:
      - description: 会话ID
        in: path
        name: session_id
        required: true
        type: string
      - description: 消息ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 删除消息
      tags:
      - 消息
  /messages/{session_id}/load:
    get:
      consumes:
      - application/json
      description: 加载会话的消息历史，支持分页和时间筛选
      parameters:
      - description: 会话ID
        in: path
        name: session_id
        required: true
        type: string
      - default: 20
        description: 返回数量
        in: query
        name: limit
        type: integer
      - description: 在此时间之前的消息（RFC3339Nano格式）
        in: query
        name: before_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 消息列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 加载消息历史
      tags:
      - 消息
  /models:
    get:
      consumes:
      - application/json
      description: 获取当前租户的所有模型
      produces:
      - application/json
      responses:
        "200":
          description: 模型列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 获取模型列表
      tags:
      - 模型管理
    post:
      consumes:
      - application/json
      description: 创建新的模型配置
      parameters:
      - description: 模型信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.CreateModelRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 创建的模型
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 创建模型
      tags:
      - 模型管理
  /models/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的模型
      parameters:
      - description: 模型ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 模型不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 删除模型
      tags:
      - 模型管理
    get:
      consumes:
      - application/json
      description: 根据ID获取模型详情
      parameters:
      - description: 模型ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 模型详情
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 模型不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 获取模型详情
      tags:
      - 模型管理
    put:
      consumes:
      - application/json
      description: 更新模型配置信息
      parameters:
      - description: 模型ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.UpdateModelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的模型
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 模型不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      - ApiKeyAuth: []
      summary: 更新模型
      tags:
      - 模型管理
...
securityDefinitions:
  ApiKeyAuth:
    description: 租户身份认证：输入 sk- 开头的 API Key
    in: header
    name: X-API-Key
    type: apiKey
  Bearer:
    description: 用户登录认证：输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"